Browser files renaming
